requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aws-wsgi==0.2.7
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
flask==3.0.3
greenlet==3.1.1
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==3.0.2
pymongo==4.10.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.1.2
